modify beforeMethod and beforeClass, create TC for KYC Valid inputs

+++ src/test/java/kpx/base/BaseTest.java
@@ -1,54 +1,57 @@
 package kpx.base;
 
-import org.mlkpx.pageObject.Google_PageObjects;
-import org.mlkpx.pageObject.Home_PageObjects;
+import mlkpx.testSteps.Home_Steps;
+import mlkpx.testSteps.Kyc_Steps;
+import mlkpx.testSteps.Login_Steps;
 import org.testng.annotations.*;
 
 import static utilities.Driver.DriverManager.*;
 import utilities.Driver.DriverType;
 
-import org.mlkpx.pageObject.Login_PageObjects;
 import utilities.Logger.LoggingUtils;
 
-public class BaseTest{
-    protected Login_PageObjects loginPage;
-    protected Google_PageObjects googlePageObjects;
-
-    protected Home_PageObjects homePageObjects;
+public class BaseTest {
+    protected Home_Steps homeSteps;
+    protected Login_Steps loginSteps;
+    protected Kyc_Steps kycSteps;
 
 
     @Parameters("browser")
     @BeforeClass (alwaysRun = true)
     public void setUp(final String browser){
-        createDriver(DriverType.valueOf(browser.toUpperCase()));
+        if (browser == null || browser.isEmpty()) {
+            throw new IllegalArgumentException("Browser parameter cannot be null or empty");
+        }
+        initializeDriver(DriverType.valueOf(browser.toUpperCase()));
         getDriver().manage().deleteAllCookies();
-
-
-    }
-
-    @BeforeMethod(alwaysRun = true)
-    public void setUpTests(){
         if (getDriver().toString().contains("RemoteWebDriver")) {
             getDriver().get(System.getProperty("targetUrl"));
         } else {
             getDriver().get(System.getProperty("targetUrl"));
+            LoggingUtils.info("Redirecting back to home");
         }
-        this.googlePageObjects = new Google_PageObjects();
-        this.loginPage = new Login_PageObjects();
-        this.homePageObjects = new Home_PageObjects();
+        this.loginSteps = new Login_Steps();
+        this.homeSteps = new Home_Steps();
+        this.kycSteps = new Kyc_Steps();
+
+    }
+    private void initializeDriver(DriverType driverType) {
+        createDriver(driverType);
+    }
+
+
+    @BeforeMethod(alwaysRun = true)
+    public void setUpTests(){
+        getDriver().get(System.getProperty("homeUrl"));
     }
     @AfterMethod(alwaysRun = true)
     public void clean(){
         LoggingUtils.info("------>>>Test Ended<<<-------");
-        LoggingUtils.info("Redirecting back to home");
-        getDriver().navigate().refresh();
-        getDriver().get(System.getProperty("targetUrl"));
 
     }
 
     @AfterClass(alwaysRun = true)
     public void tearDown () {
-        quitDriver();
-
+        closeWebBrowser();
     }
 }
